Refactor LLM client response handling with unified stream and parsing logic
This commit introduces a comprehensive refactoring of LLM client response handling across different providers (OpenAI, Anthropic, Google, Vertex) by:

- Adding a new `stream_request.rs` module to centralize streaming response logic
- Creating response handler modules for each provider with consistent parsing and streaming methods
- Updating client implementations to use the new centralized response handling
- Introducing a `ResponseType` enum to dynamically route response parsing
- Simplifying stream and non-stream request processing in client methods
