+++ engine/baml-runtime/src/internal/llm_client/primitive/anthropic/mod.rs
@@ -1,4 +1,5 @@
 mod anthropic_client;
 mod types;
+pub mod response_handler;
 
 pub use anthropic_client::AnthropicClient;
