+++ engine/baml-runtime/src/internal/llm_client/primitive/google/mod.rs
@@ -1,4 +1,5 @@
 mod googleai_client;
 mod types;
+pub(crate) mod response_handler;
 
 pub use googleai_client::GoogleAIClient;
