+++ engine/baml-runtime/src/internal/llm_client/primitive/openai/mod.rs
@@ -1,5 +1,6 @@
 mod openai_client;
 mod properties;
+pub mod response_handler;
 #[allow(dead_code)]
 mod types;
