+++ engine/baml-runtime/src/internal/llm_client/primitive/vertex/mod.rs
@@ -1,5 +1,6 @@
 #[cfg(target_arch = "wasm32")]
 pub(super) mod wasm_auth;
+pub(crate) mod response_handler;
 #[cfg(target_arch = "wasm32")]
 pub(super) use wasm_auth as auth;
