+++ engine/baml-runtime/src/internal/llm_client/primitive/vertex/vertex_client.rs
@@ -1,4 +1,6 @@
 use crate::client_registry::ClientProperty;
+use crate::internal::llm_client::primitive::request::ResponseType;
+use crate::internal::llm_client::primitive::stream_request::make_stream_request;
 use crate::internal::llm_client::traits::{
     ToProviderMessage, ToProviderMessageExt, WithClientProperties,
 };
@@ -108,95 +110,6 @@ impl WithClient for VertexClient {
 
 impl WithNoCompletion for VertexClient {}
 
-impl SseResponseTrait for VertexClient {
-    fn response_stream(
-        &self,
-        resp: reqwest::Response,
-        prompt: &[RenderedChatMessage],
-        system_start: web_time::SystemTime,
-        instant_start: web_time::Instant,
-    ) -> StreamResponse {
-        let prompt = prompt.to_vec();
-        let client_name = self.context.name.clone();
-        let model_id = self.properties.model.clone();
-        let params = self.properties.properties.clone();
-        Ok(Box::pin(
-            resp.bytes_stream()
-                .eventsource()
-                .inspect(|event| log::trace!("Received event: {:?}", event))
-                .take_while(|event| {
-                    std::future::ready(event.as_ref().is_ok_and(|e| e.data != "data: \n"))
-                })
-                .map(|event| -> Result<VertexResponse> {
-                    Ok(serde_json::from_str::<VertexResponse>(&event?.data)?)
-                })
-                .scan(
-                    Ok(LLMCompleteResponse {
-                        client: client_name.clone(),
-                        prompt: internal_baml_jinja::RenderedPrompt::Chat(prompt.clone()),
-                        content: "".to_string(),
-                        start_time: system_start,
-                        latency: instant_start.elapsed(),
-                        model: model_id,
-                        request_options: params.clone(),
-                        metadata: LLMCompleteResponseMetadata {
-                            baml_is_complete: false,
-                            finish_reason: None,
-                            prompt_tokens: None,
-                            output_tokens: None,
-                            total_tokens: None,
-                        },
-                    }),
-                    move |accumulated: &mut Result<LLMCompleteResponse>, event| {
-                        let Ok(ref mut inner) = accumulated else {
-                            // halt the stream: the last stream event failed to parse
-                            return std::future::ready(None);
-                        };
-                        let event = match event {
-                            Ok(event) => event,
-                            Err(e) => {
-                                return std::future::ready(Some(LLMResponse::LLMFailure(
-                                    LLMErrorResponse {
-                                        client: client_name.clone(),
-                                        model: if inner.model.is_empty() {
-                                            None
-                                        } else {
-                                            Some(inner.model.clone())
-                                        },
-                                        prompt: internal_baml_jinja::RenderedPrompt::Chat(
-                                            prompt.to_vec(),
-                                        ),
-                                        start_time: system_start,
-                                        request_options: params.clone(),
-                                        latency: instant_start.elapsed(),
-                                        message: format!("Failed to parse event: {:#?}", e),
-                                        code: ErrorCode::UnsupportedResponse(2),
-                                    },
-                                )));
-                            }
-                        };
-                        if let Some(choice) = event.candidates.first() {
-                            if let Some(content) = choice
-                                .content
-                                .as_ref()
-                                .and_then(|c| c.parts.first().map(|p| p.text.as_ref()))
-                            {
-                                inner.content += content;
-                            }
-                            if let Some(FinishReason::Stop) = choice.finish_reason.as_ref() {
-                                inner.metadata.baml_is_complete = true;
-                                inner.metadata.finish_reason = Some(FinishReason::Stop.to_string());
-                            }
-                        }
-
-                        inner.latency = instant_start.elapsed();
-
-                        std::future::ready(Some(LLMResponse::Success(inner.clone())))
-                    },
-                ),
-        ))
-    }
-}
 // makes the request to the google client, on success it triggers the response_stream function to handle continuous rendering with the response object
 impl WithStreamChat for VertexClient {
     async fn stream_chat(
@@ -205,12 +118,7 @@ impl WithStreamChat for VertexClient {
         prompt: &[RenderedChatMessage],
     ) -> StreamResponse {
         //incomplete, streaming response object is returned
-        let (response, system_now, instant_now) =
-            match make_request(self, either::Either::Right(prompt), true).await {
-                Ok(v) => v,
-                Err(e) => return Err(e),
-            };
-        self.response_stream(response, prompt, system_now, instant_now)
+        make_stream_request(self, either::Either::Right(prompt), Some(self.properties.model.clone()), ResponseType::Vertex).await
     }
 }
 
@@ -298,6 +206,8 @@ impl RequestBuilder for VertexClient {
         let baml_original_url = format!(
             "{base_url}/{model}:{rpc_and_protocol}",
             model = self.properties.model,
+            // publisher "anthropic" "google"
+            // rawPredict streamRawPredict
             rpc_and_protocol = if stream {
                 "streamGenerateContent?alt=sse"
             } else {
@@ -343,79 +253,16 @@ impl RequestBuilder for VertexClient {
 
 impl WithChat for VertexClient {
     async fn chat(&self, _ctx: &RuntimeContext, prompt: &[RenderedChatMessage]) -> LLMResponse {
+        let model_name = self.properties.model.clone();
         //non-streaming, complete response is returned
-        let (response, system_now, instant_now) =
-            match make_parsed_request::<VertexResponse>(self, either::Either::Right(prompt), false)
-                .await
-            {
-                Ok(v) => v,
-                Err(e) => return e,
-            };
-
-        if response.candidates.len() != 1 {
-            return LLMResponse::LLMFailure(LLMErrorResponse {
-                client: self.context.name.to_string(),
-                model: None,
-                prompt: internal_baml_jinja::RenderedPrompt::Chat(prompt.to_vec()),
-                start_time: system_now,
-                request_options: self.properties.properties.clone(),
-                latency: instant_now.elapsed(),
-                message: format!(
-                    "Expected exactly one content block, got {}",
-                    response.candidates.len()
-                ),
-                code: ErrorCode::Other(200),
-            });
-        }
-
-        let content = if let Some(content) = response.candidates.first().and_then(|c| {
-            c.content
-                .as_ref()
-                .and_then(|c| c.parts.first().map(|p| p.text.clone()))
-        }) {
-            content
-        } else {
-            return LLMResponse::LLMFailure(LLMErrorResponse {
-                client: self.context.name.to_string(),
-                model: None,
-                prompt: internal_baml_jinja::RenderedPrompt::Chat(prompt.to_vec()),
-                start_time: system_now,
-                request_options: self.properties.properties.clone(),
-                latency: instant_now.elapsed(),
-                message: "No content".to_string(),
-                code: ErrorCode::Other(200),
-            });
-        };
-
-        let usage_metadata = response.usage_metadata.clone().unwrap();
-
-        LLMResponse::Success(LLMCompleteResponse {
-            client: self.context.name.to_string(),
-            prompt: internal_baml_jinja::RenderedPrompt::Chat(prompt.to_vec()),
-            content,
-            start_time: system_now,
-            latency: instant_now.elapsed(),
-            request_options: self.properties.properties.clone(),
-            model: self
-                .properties
-                .properties
-                .get("model")
-                .and_then(|v| v.as_str().map(|s| s.to_string()))
-                .unwrap_or_default(),
-            metadata: LLMCompleteResponseMetadata {
-                baml_is_complete: matches!(
-                    response.candidates[0].finish_reason,
-                    Some(FinishReason::Stop)
-                ),
-                finish_reason: response.candidates[0]
-                    .finish_reason
-                    .as_ref()
-                    .map(|r| serde_json::to_string(r).unwrap_or("".into())),
-                prompt_tokens: usage_metadata.prompt_token_count,
-                output_tokens: usage_metadata.candidates_token_count,
-                total_tokens: usage_metadata.total_token_count,
-            },
-        })
+        make_parsed_request(
+            self,
+            Some(model_name),
+            either::Either::Right(prompt),
+            false,
+            ResponseType::Vertex,
+        )
+        .await
     }
 }
